% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_clauses.R
\name{add_sql_returning}
\alias{add_sql_returning}
\title{Add a returning clause}
\usage{
add_sql_returning(sql, returning, con)
}
\arguments{
\item{sql}{An SQL clause.}

\item{returning}{An \href{sql-expression-list}{SQL expression list} that specifies
which columns to return. Names are optional and used as column names in
the returned table. For the default \code{NULL} no rows are returned.}

\item{con}{A DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}
}
\description{
Add a returning clause
}
