% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.numeric_noDataLost}
\alias{as.numeric_noDataLost}
\title{Change data type to numeric if no data is lost}
\usage{
as.numeric_noDataLost(x, warning = FALSE)
}
\arguments{
\item{x}{a variable of type \code{data.frame}, \code{list} or atomic.}

\item{warning}{if `TRUE`, throw a warning indicating which values can not be casted to numeric.}
}
\value{
returns the same variable as a numeric type only if there is no data lost. For \code{data.frame} and \code{list},
 it change the data type per column/element.
}
\description{
Change data type to numeric if no data is lost
}
