% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{auto_name}
\alias{auto_name}
\title{Auto name a list or character vector}
\usage{
auto_name(x)
}
\description{
Auto name a list or character vector
}
\examples{
auto_name(list("body"))
list("body", left = "right", SQL("unnamed_sql"), left = SQL("named_sql")) \%>\%
  auto_name()
}
