% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoolYear.R
\name{BedsDate}
\alias{BedsDate}
\title{BEDS Date}
\usage{
BedsDate(year = schoolYear())
}
\arguments{
\item{year}{integer - the 4 digit year at the beginning of the school year.  Defaults to current school year.}
}
\value{
date - the date of BEDS day for the given school year
}
\description{
Determine the date of BEDS day in a school year (first Wednesday in October)
}
\examples{
BedsDate()
BedsDate(2013)
}
