% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterGrep.R
\name{betterGrep}
\alias{betterGrep}
\title{Better grep}
\usage{
betterGrep(pattern, x, ignore.case = T, fixed = T)
}
\arguments{
\item{pattern}{character string containing a regular expression (or character string for fixed = TRUE) to be matched in the given character vector. Coerced by as.character to a character string if possible. If a character vector of length 2 or more is supplied, the first element is used with a warning.}

\item{x}{a character vector where matches are sought, or an object which can be coerced by as.character to a character vector. Long vectors are supported.}

\item{ignore.case}{logical.  if FALSE, the pattern matching is case sensitive and if TRUE, case is ignored during matching.}

\item{fixed}{logical. If TRUE, pattern is a string to be matched as is. Overrides all conflicting arguments.}
}
\value{
Numeric vector holding the indices of elements of \code{x} that had matches
}
\description{
Search for pattern matches, where trying to match to NA comes back as FALSE
}
\examples{
x = c("Alpha", "Beta", "Bet", NA_character_, "B")
betterGrep(pattern = "a", x)
betterGrep(pattern = NA, x)
betterGrep(pattern = NA, x, ignore.case = F)
betterGrep("Beta", x)
betterGrep("B", x)
betterGrep("X", x)
}
