% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterGrepl.R
\name{betterGrepl.any}
\alias{betterGrepl.any}
\title{Better grepl any}
\usage{
betterGrepl.any(pattern, x, ignore.case = T, fixed = T)
}
\arguments{
\item{pattern}{character string containing a regular expression (or character string for fixed = TRUE) to be matched in the
given character vector. Coerced by as.character to a character string if possible. If a character vector of length 2 or more
is supplied, the first element is used with a warning.}

\item{x}{a character vector where matches are sought, or an object which can be coerced by as.character to a character vector.
Long vectors are supported.}

\item{ignore.case}{logical.  if FALSE, the pattern matching is case sensitive and if TRUE, case is ignored during matching.}

\item{fixed}{logical. If TRUE, pattern is a string to be matched as is. Overrides all conflicting arguments.}
}
\value{
Logical vector indicating whether any elements had matches
}
\description{
This function behaves similarly to betterGrepl, except that it returns only one logical value
}
\examples{
x = c("Alpha", "Beta", "Bet", NA_character_, "B")
betterGrepl.any(pattern = "a", x)
betterGrepl.any(pattern = NA, x)
betterGrepl.any(pattern = NA, x, ignore.case = F)
betterGrepl.any("Beta", x)
betterGrepl.any("beta", x)
betterGrepl.any("beta", x, ignore.case = F)
betterGrepl.any("B", x)
betterGrepl.any("X", x)
}
