% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterGrepl.R
\name{betterGrepl}
\alias{betterGrepl}
\title{Better grepl}
\usage{
betterGrepl(pattern, x, ignore.case = T, fixed = T)
}
\arguments{
\item{pattern}{character string containing a regular expression (or character string for fixed = TRUE) to be matched in the given character vector. Coerced by as.character to a character string if possible. If a character vector of length 2 or more is supplied, the first element is used with a warning.}

\item{x}{a character vector where matches are sought, or an object which can be coerced by as.character to a character vector. Long vectors are supported.}

\item{ignore.case}{logical.  if FALSE, the pattern matching is case sensitive and if TRUE, case is ignored during matching.}

\item{fixed}{logical. If TRUE, pattern is a string to be matched as is. Overrides all conflicting arguments.}
}
\value{
Logical vector of the same length as \code{x} indicating which elements had matches
}
\description{
This function behaves similarly to grepl, except that an NA pattern or NA element of \code{x} returns FALSE, but NA matched to NA returns TRUE
}
\examples{
x = c("Alpha", "Beta", "Bet", NA_character_, "B")
betterGrepl(pattern = "a", x)
betterGrepl(pattern = NA, x)
betterGrepl(pattern = NA, x, ignore.case = F)
betterGrepl("Beta", x)
betterGrepl("B", x)
betterGrepl("X", x)
}
