% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterMax.R
\name{betterMax}
\alias{betterMax}
\title{Find the maximum value of a set that may contain NA's.}
\usage{
betterMax(x)
}
\arguments{
\item{x}{A numerical vector}
}
\value{
The largest value in \code{x}, if any.  Otherwise, NA.
}
\description{
Find the maximum value of a set that may contain NA's.
}
\examples{
betterMax(1:5)
betterMax(c(2,3,NA,7))
betterMax(c(NA, NA, NA))
}
