% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterMin.R
\name{betterMin}
\alias{betterMin}
\title{Better Min}
\usage{
betterMin(x)
}
\arguments{
\item{x}{An object such as a numeric vector, matrix, or data.frame}
}
\value{
The largest value in \code{x}, if any.  Otherwise, NA.
}
\description{
Find the minimum value of a object that may contain NA's.
}
\examples{
betterMin(1:5)
betterMin(c(2,3,NA,7))
betterMin(c(NA, NA, NA))
ex1 = data.frame(X1 = LETTERS, X2 = 1:26, X3 = sample(c(20,NA), 26, T))
betterMin(ex1)
}
