% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharCartProd.R
\name{CharCartProd}
\alias{CharCartProd}
\title{Character Cartesian Product}
\usage{
CharCartProd(x, y)
}
\arguments{
\item{x}{A character vector}

\item{y}{A character vector}
}
\value{
A character vector where each element is a concatenation of one element from x and one element from y
}
\description{
Create a character vector of the cartesian product of two character vectors
}
\examples{
vecOne = letters[1:5]
vecTwo = LETTERS[6:8]
CharCartProd(x = vecOne, y = vecTwo)
CharCartProd(vecTwo, vecOne)
}
