% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr_compat.R
\name{con_memdb}
\alias{con_memdb}
\alias{memdb_frame2}
\title{Create a database table in temporary in-memory database}
\usage{
con_memdb()

memdb_frame2(..., .name)
}
\arguments{
\item{...}{Passed to \code{\link[tibble:tibble]{tibble::tibble()}}.}

\item{.name}{Name of table in database.}
}
\description{
\code{memdb_frame2()} works like \code{\link[dbplyr:memdb_frame]{dbplyr::memdb_frame()}}: it creates a table in
an in-memory SQLite database.
}
