% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_conflict_helpers.R
\name{conflict-target}
\alias{conflict-target}
\alias{sql_constraint}
\alias{sql_unique_cols}
\title{SQL conflict target}
\usage{
sql_constraint(constraint)

sql_unique_cols(...)
}
\arguments{
\item{constraint}{The name of a constraint (a scalar character). The
constraint must be allowed to in use in the \verb{ON CONFLICT ON CONSTRAINT}
clause.}

\item{...}{A character vector of names of unique columns.}
}
\description{
SQL conflict target
}
\examples{
sql_constraint("unique_index")

sql_unique_cols("a", "b")
}
