% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyClipboard.R
\name{CopyClipboard}
\alias{CopyClipboard}
\title{Copy Clipboard}
\usage{
CopyClipboard(allowEscapes = F, strip.white = F, sep = ">",
  single = T)
}
\arguments{
\item{allowEscapes}{Logical, should escape characters be interpreted?}

\item{strip.white}{Logical, should whitespace be stripped from the beginning and end of the clipboard contents?}

\item{sep}{Single character by which to split the string}

\item{single}{Logical, should the clipboard contents be returned as a single string, no matter what?}
}
\value{
Character vector with the clipboard contents
}
\description{
Copy text from the clipboard as-is
}
