% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{create_temp_table}
\alias{create_temp_table}
\title{Create a temporary table}
\usage{
create_temp_table(sql, table_name)
}
\arguments{
\item{table_name}{The name of the temp table you wish to create}

\item{sql:}{The SQL table you want to create a temp table out of. Should
be a table that starts with a WITH or SELECT clause.}
}
\description{
Create a temporary table
}
\examples{
`dbtools::create_temp_table("SELECT a_col, count(*) as n FROM a_database.table GROUP BY a_col", table_name="temp_table_1")`
}
