% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCompareMatrix.R
\name{CreateCompareMatrix}
\alias{CreateCompareMatrix}
\title{Create Comparison Matrix}
\usage{
CreateCompareMatrix(studentlist, Exams, matrix1, matrix2)
}
\arguments{
\item{studentlist}{A vector of row names}

\item{Exams}{A vector of column names}

\item{matrix1}{A numeric matrix}

\item{matrix2}{A numeric matrix with the same dimensions as  \code{matrix1}}
}
\value{
A numeric matrix with the same dimensions as the input matrices
}
\description{
Given two numeric matrices with equal dimensions, create a matrix of max values of corresponding cells and apply row and column names.  NA's are only returned if both matrices have NA in the same cell.
}
\examples{
s1 = LETTERS[1:5]
E1 = c("Math", "Science", "History")
m1 = matrix(sample(100, 15, F), 5)
m2 = matrix(c(NA, 7*(1:14)), 5)
m3 = matrix(c(NA, NA, 100 - 7*(1:13)), 5)
CreateCompareMatrix(s1, E1, m1, m2)
CreateCompareMatrix(s1, E1, m1, m3)
CreateCompareMatrix(s1, E1, m2, m3)
}
