% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Credentials.R
\docType{class}
\name{Credentials-class}
\alias{Credentials}
\alias{Credentials-class}
\alias{as.list.Credentials}
\title{Credentials Class}
\usage{
Credentials(drv, ...)

\method{as.list}{Credentials}(x, ...)
}
\arguments{
\item{drv}{(function) a driver. Will be called and passed to \link[DBI]{dbConnect}}

\item{...}{named arguments which can be passed to \link[DBI]{dbConnect}}

\item{x}{(Credentials) an instance}
}
\description{
Use this function to store your database credentials.
}

