% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-column.R
\name{db_column}
\alias{db_column}
\alias{new_db_column}
\title{Database Column}
\usage{
new_db_column(
  x = list(),
  name,
  default = NULL,
  validate = function(value) { },
  nullable = TRUE,
  ...,
  class = character()
)
}
\arguments{
\item{x}{base type (list)}

\item{name}{name of the column, as string}

\item{default}{default value of the column in expression}

\item{validate}{function that takes the value as an argument, returns
TRUE or NULL if the provided data is valid, throws an error
or returns anything other than a TRUE value.}

\item{nullable}{whether or not this column is nullable, default TRUE}

\item{...}{extra attributes}

\item{class}{sub-classes as character string}
}
\description{
These constructors represent database columns and allow us to specify the
structure of a database table.
}
