% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-column-boolean.R
\name{db_column_boolean}
\alias{db_column_boolean}
\title{Boolean Column}
\usage{
db_column_boolean(
  table,
  name,
  default = NULL,
  validate = function(value) { },
  nullable = TRUE
)
}
\description{
Boolean Column
}
