% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-column-date.R
\name{db_column_date}
\alias{db_column_date}
\title{Date Column}
\usage{
db_column_date(
  table,
  name,
  default = NULL,
  validate = function(value) { },
  nullable = TRUE
)
}
\description{
Date Column
}
