% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-column-datetime.R
\name{db_column_datetime}
\alias{db_column_datetime}
\title{Date-Time (Timestamp) Column}
\usage{
db_column_datetime(
  name,
  default = NULL,
  validate = function(value) { },
  nullable = TRUE,
  .x = list(),
  .class = character()
)
}
\description{
Date-Time (Timestamp) Column
}
