% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-column-double.R
\name{db_column_double}
\alias{db_column_double}
\title{Double Precision Column}
\usage{
db_column_double(
  name,
  default = NULL,
  validate = function(value) { },
  nullable = TRUE,
  .x = list(),
  .class = character()
)
}
\description{
Double Precision Column
}
