% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-column-integer.R
\name{db_column_integer}
\alias{db_column_integer}
\title{Integer Column}
\usage{
db_column_integer(
  table,
  name,
  default = NULL,
  int_type = c("integer", "smallint", "bigint"),
  validate = function(value) { },
  nullable = TRUE,
  identity = c("none", "by default", "always")
)
}
\arguments{
\item{identity}{is this an identity column? Can be:
\itemize{
\item \code{"none"}    - not an identity column
\item \code{"default"} - can be overwritten on insert
\item \code{"always"}  - can not be overwritten
}}
}
\description{
Integer Column
}
