% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-column-json.R
\name{db_column_json}
\alias{db_column_json}
\title{JSON Column}
\usage{
db_column_json(
  table,
  name,
  binary = TRUE,
  default = NULL,
  validate = function(value) { },
  nullable = TRUE
)
}
\description{
JSON Column
}
