% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-column-uuid.R
\name{db_column_uuid}
\alias{db_column_uuid}
\title{UUID Column}
\usage{
db_column_uuid(
  table,
  name,
  default = NULL,
  validate = function(value) { },
  nullable = TRUE
)
}
\description{
A Universally Unique Identifier column
}
