% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-constraint.R
\name{db_constraint}
\alias{db_constraint}
\alias{new_db_constraint}
\title{Database Constraint}
\usage{
db_constraint(table, constraint)

new_db_constraint(x = list(), name, ..., class = character())
}
\arguments{
\item{name}{constraint name}
}
\description{
Database Constraint
}
