% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-constraint-check.R
\name{db_constraint_check}
\alias{db_constraint_check}
\title{Primary Key Constraint}
\usage{
db_constraint_check(table, check, name = NULL)
}
\arguments{
\item{table}{database table}

\item{check}{check expression to be checked against new inputs}

\item{name}{name of the constraint}
}
\description{
Primary Key Constraint
}
