% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-constraint-fkey.R
\name{db_constraint_fkey}
\alias{db_constraint_fkey}
\title{Foreign Key Constraint}
\usage{
db_constraint_fkey(
  table,
  foreign_table,
  column,
  foreign_column = NULL,
  on_update = c("no action", "set null", "restrict", "set default", "cascade"),
  on_delete = c("no action", "set null", "restrict", "set default", "cascade"),
  name = NULL
)
}
\arguments{
\item{table}{database table}

\item{foreign_table}{foreign database table}

\item{column}{table column to be used}

\item{foreign_column}{foreign column}

\item{on_update}{setting for when the foreign key is updated}

\item{on_delete}{setting for when the foreign key is deleted}

\item{name}{name of the constraint}
}
\description{
Foreign Key Constraint
}
