% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-constraint-pkey.R
\name{db_constraint_pkey}
\alias{db_constraint_pkey}
\title{Primary Key Constraint}
\usage{
db_constraint_pkey(table, ..., name = NULL)
}
\arguments{
\item{table}{database table}

\item{...}{unquoted column_names subject to the primary key constraint}

\item{name}{name of the constraint}
}
\description{
Primary Key Constraint
}
