% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-index.R
\name{db_index}
\alias{db_index}
\alias{new_db_index}
\title{Database Index}
\usage{
db_index(table, index)

new_db_index(
  x = list(),
  table,
  name,
  ...,
  unique = FALSE,
  partial = NULL,
  class = character()
)
}
\arguments{
\item{name}{index name}

\item{unique}{whether or not index must be on unique values}

\item{partial}{if partial index, R expression that defines the bounds of the
index, but NULL if not partial}
}
\description{
Database Index
}
