% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-index-column.R
\name{db_index_column}
\alias{db_index_column}
\title{Database Index based on a column or columns}
\usage{
db_index_column(table, ..., unique = FALSE, partial = NULL, name = NULL)
}
\arguments{
\item{...}{unquoted column names subject to the index}
}
\description{
Database Index based on a column or columns
}
