% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-index-expression.R
\name{db_index_expression}
\alias{db_index_expression}
\title{Database Index based on a SQL expression}
\usage{
db_index_expression(
  table,
  expression,
  unique = FALSE,
  partial = NULL,
  name = NULL
)
}
\arguments{
\item{table}{database table}

\item{expression}{unquoted expression that will form the index}

\item{name}{name of the index}
}
\description{
Database Index based on a SQL expression
}
