% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils_table_size.R
\name{db_utils_index_infos}
\alias{db_utils_index_infos}
\title{Get database index infos}
\usage{
db_utils_index_infos(con, table = NULL)
}
\arguments{
\item{con}{A DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{table}{The name of the database table to update, insert to, or upsert
to. Can be an unquoted character string or an identifier created with \code{\link[=ident]{ident()}}.}
}
\description{
Get database index infos
}
