% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils_table_size.R
\name{db_utils_running_queries}
\alias{db_utils_running_queries}
\title{Get running queries}
\usage{
db_utils_running_queries(con)
}
\arguments{
\item{con}{a DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}
}
\description{
Get running queries
}
