% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils_table_size.R
\name{db_utils_table_size}
\alias{db_utils_table_size}
\title{Get database table sizes}
\usage{
db_utils_table_size(con)
}
\arguments{
\item{con}{A DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}
}
\description{
Get database table sizes
}
