% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DBCredentials.R
\name{DBCredentials}
\alias{DBCredentials}
\title{DBCredentials constructor}
\usage{
DBCredentials(...)
}
\arguments{
\item{...}{arguments which can be passed to \code{dbConnect}}
}
\description{
Funktion can be used to construct a new credentials object
}

