% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbt_hlp_progress.R
\name{dbt_hlp_progress}
\alias{dbt_hlp_progress}
\title{dbt_hlp_progress}
\usage{
dbt_hlp_progress(i = NULL, ii = NULL, start = NULL, redraw = FALSE,
  now = Sys.time())
}
\arguments{
\item{i}{ith iteration / item / job / query}

\item{ii}{number of iterations / ... in total}

\item{start}{start time}
}
\description{
dbt_hlp_progress
}
\examples{

dbt_hlp_progress(1)
dbt_hlp_progress(3, 10, Sys.time()-10)
}
