% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbt_sql.R
\name{dbt_sql}
\alias{dbt_sql}
\title{dbt_sql}
\usage{
dbt_sql(query_string, ...)
}
\arguments{
\item{query_string}{a (character) vector or SQL to be either marked as SQL or
to be interpolated.}

\item{...}{optional furtherarguments used to interpolate values into
query_string}
}
\value{
an object of type SQL
}
\description{
dbt_sql
}
