% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{dbtEnvExists}
\alias{dbtEnvExists}
\title{Check whether .dbcon environment (for storing connections) exists}
\usage{
dbtEnvExists()
}
\description{
Check whether .dbcon environment (for storing connections) exists
}
