% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_insert.R
\name{default-args}
\alias{default-args}
\title{Common arguments}
\arguments{
\item{data}{a data.frame.}

\item{table}{name of the database table.}

\item{con}{a DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}
}
\description{
Common arguments
}
