% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{delete_database_and_data}
\alias{delete_database_and_data}
\title{Deletes both an Athena database and the underlying data on S3.}
\usage{
delete_database_and_data(database)
}
\arguments{
\item{database}{Database name}
}
\description{
Deletes both an Athena database and the underlying data on S3.
}
