% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delete_object}
\alias{delete_object}
\title{delete_object}
\usage{
delete_object(bucket, key)
}
\arguments{
\item{bucket}{A string specifying the s3 bucket name}

\item{key}{File path to the s3 object}
}
\description{
uses boto3 (in python) to delete an S3 object (used by read_sql function to clean up after itself)
}
\details{
Will send an SQL query to athena and wait for it to complete. Once the query has completed the resulting sql query will be read using read.csv (base R), read_csv (readr) or fread (data.table).
Function returns dataframe. If needing more a more bespoke or self defined data reading function and arguments use dbtools::get_athena_query_response to send an SQL query and return the s3 path to data in csv format.
}
\examples{
# delete a file from S3
# (note this is not exported so have to use tripple colon)
dbtools:::delete_object('my_bucket', 'path/to/file.csv')
}
