% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{delete_partitions_and_data}
\alias{delete_partitions_and_data}
\title{Deletes partitions and the underlying data on S3 from an Athena
database table matching an expression.}
\usage{
delete_partitions_and_data(database, table, expression)
}
\arguments{
\item{database}{The database name.}

\item{table}{The table name.}

\item{expression}{The expression to match.}
}
\description{
Deletes partitions and the underlying data on S3 from an Athena
database table matching an expression.
}
\examples{
`dbtools::delete_partitions_and_data("my_database", "my_table", "year = 2020 and month = 5")`
}
