% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{delete_table_and_data}
\alias{delete_table_and_data}
\title{Deletes both a table from an Athena database and the underlying data on S3.}
\usage{
delete_table_and_data(database, table)
}
\arguments{
\item{database}{The database name.}

\item{table}{The table name.}
}
\description{
Deletes both a table from an Athena database and the underlying data on S3.
}
\examples{
`dbtools::delete_table_and_data("__temp__", "my_table")`
}
