% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{describe_table}
\alias{describe_table}
\title{Show the list of columns, including partition columns: 'DESCRIBE table;'.}
\usage{
describe_table(table, database)
}
\arguments{
\item{table}{Table name}

\item{database}{AWS Glue/Athena database name}
}
\value{
DataFrame filled by formatted infos.
}
\description{
Show the list of columns, including partition columns: 'DESCRIBE table;'.
}
\examples{
`df_table = dbtools::describe_table(table='my_table', database='my_db')`
}
