% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleSetDiff.R
\name{doubleSetDiff}
\alias{doubleSetDiff}
\title{Double Set Difference}
\usage{
doubleSetDiff(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{A vector of the same type as /code{x}}
}
\value{
a vector containing all of the unique elements that are members of x or y, but not of both
}
\description{
Compare two sets and return a vector of all items that appear in exactly one of them
}
