% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primaryKey.R
\name{duplicatedPK}
\alias{duplicatedPK}
\title{Find registers with a duplicated Primary Key}
\usage{
duplicatedPK(x, pk)
}
\arguments{
\item{x}{a \code{data.frame}}

\item{pk}{columns containing the Primary Key (column names or indexes)}
}
\value{
a \code{data.frame} with the duplicated registers or \code{NA} if there are no duplicated PK values after \code{unique(x)}.
}
\description{
Find registers with a duplicated Primary Key
}
