% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindInconsistentActiveStatus.R
\name{FindInconsistentActiveStatus}
\alias{FindInconsistentActiveStatus}
\title{Find Inconsistent Active Status}
\usage{
FindInconsistentActiveStatus(psStudentsRaw, Workbook)
}
\arguments{
\item{psStudentsRaw}{an export from the students table in PowerSchool}

\item{Workbook}{a data.frame of the All Info tab from the workbook}
}
\value{
either a dataframe of issues or numeric(0) if there are no issues
}
\description{
Identify students who have different statuses in PowerSchool and the Workbook
}
