% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_athena_connection.R
\name{get_athena_connection}
\alias{get_athena_connection}
\title{get_athena_connection}
\usage{
get_athena_connection()
}
\value{
A connection to the databases on AWS
}
\description{
establishes a connection to the databases on AWS allowing users to query the data with SQL
}
\details{
Returns a connection to all the databases stored on AWS. You make be able to see all the database but you may not be able to query them depending on what permissions
you have to S3 buckets. If you get an access denied error when trying to query a database please contact the owner of the bucket in which the data exists and ask for access to the
data.
}
\examples{
# Get the connection to the crest data and then query crest using get_athena_connection

bucket <- alpha-dag-crest-data-engineering
con <- get_athena_connection(bucket)
data <- RJDBC::dbGetQuery(con, 'SELECT * FROM crest_v1.flatfile limit 1000')
}
