% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{get_query_columns_types}
\alias{get_query_columns_types}
\title{Get the data type of all columns queried.}
\usage{
get_query_columns_types(query_id)
}
\arguments{
\item{query_id}{Athena query execution ID}
}
\value{
List with all data types
}
\description{
Get the data type of all columns queried.
}
\examples{
`dbtools::get_query_columns_types('query-execution-id')`
}
