% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{get_sql_from_file}
\alias{get_sql_from_file}
\title{Read in an SQL file and inject arguments with Jinja (if given params).}
\usage{
get_sql_from_file(filepath, jinja_args = NULL)
}
\arguments{
\item{filepath}{A filepath to your SQL file.}

\item{jinja_args}{If not NULL, will pass the read
in SQL file through a jinja template to render the template.
Otherwise will just return the SQL file as is. Defaults to NULL.}
}
\value{

}
\description{
Read in an SQL file and inject arguments with Jinja (if given params).
}
