% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNiceColumnNames.R
\name{GetNiceColumnNames}
\alias{GetNiceColumnNames}
\title{Get Nice Column Names}
\usage{
GetNiceColumnNames(sheet, templates)
}
\arguments{
\item{sheet}{character vector of length 1 with the exact name of a sheet in the template workbook}

\item{templates}{an openxlsx workbook object hold the template workbook}
}
\value{
A character vector with the names of the columns in a SIRS export with all of the non-alphanumeric characters removed
}
\description{
Get the column names of a particular SIRS export
}
