% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geocode.R
\name{getSchoolDistrict}
\alias{getSchoolDistrict}
\title{Get School District}
\usage{
getSchoolDistrict(addr, halt = T, messageLevel = 0)
}
\arguments{
\item{addr}{Character vector of length 1 containing an address}

\item{halt}{logical of length 1 indicating whether to halt if no school district is found}

\item{messageLevel}{integer of length 1 indicating the level of messaging to print}
}
\value{
Character vector of length 1 containing the name of the school district or NA
}
\description{
Takes a one line address (USA only) and returns the school district serving that address using the first match from the US Census API
}
\examples{
getSchoolDistrict("1 MAIN ST, ALBANY, NY, 12204")
getSchoolDistrict("1234 MAIN Blvd, ALBANY, NY, 12204")
getSchoolDistrict("1234 MAIN Blvd, ALBANY, NY, 12204", halt = F)
}
