% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbt_sql.R, R/sql_plus_operator.R,
%   R/sql_values.R
\docType{methods}
\name{dbt_sql,ANY-method}
\alias{dbt_sql,ANY-method}
\alias{dbt_sql,SQL-method}
\alias{dbt_sql,missing-method}
\alias{+,SQL,ANY-method}
\alias{hidden_aliases}
\alias{sql_values,ANY,ANY,missingOrNULL-method}
\alias{sql_values,vector,ANY,DBIConnection-method}
\alias{sql_values,data.frame,ANY,DBIConnection-method}
\title{hidden aliases}
\usage{
\S4method{dbt_sql}{ANY}(query_string, ...)

\S4method{dbt_sql}{SQL}(query_string, ...)

\S4method{dbt_sql}{missing}(query_string, ...)

\S4method{+}{SQL,ANY}(e1, e2)

\S4method{sql_values}{ANY,ANY,missingOrNULL}(x, ..., val_sep = "\\n",
  conn = NULL)

\S4method{sql_values}{vector,ANY,DBIConnection}(x, ..., val_sep = "\\n",
  conn = NULL)

\S4method{sql_values}{data.frame,ANY,DBIConnection}(x, ...,
  val_sep = "\\n", conn = NULL)
}
\description{
hidden aliases
}
