% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MessageLevel.R
\name{inc}
\alias{inc}
\title{inc}
\usage{
inc(ML)
}
\arguments{
\item{ML}{integer vector length 2.  The first element indicates how deep to print messages.  The second element indicates how
many layers deep the processing is when the call occurs.}
}
\value{
ML is returned, but with the second element increased by 1
}
\description{
Increment the message level
}
\note{
This function is used when ML is passed as an argment to another function.  That way, in the called function, the depth
  is 1 more than in the calling environment.
}
