% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShortFunctions.R
\name{intersectSeveral}
\alias{intersectSeveral}
\title{Intersect Several}
\usage{
intersectSeveral(...)
}
\arguments{
\item{...}{Any number of vectors of the same mode}
}
\value{
A vector of the same mode as the inputs containing elements that appear in all of the input sets.  If there are none, it will return an empty vector.
}
\description{
Find intersection of several sets
}
\examples{
ex1 = 6:10
ex2 = (1:5)*3
ex3 = c(1, 10, NA, 6, 2)
ex4 = c(NA, NA, NA, 6, 5)
intersectSeveral(ex1, ex2)
intersectSeveral(ex1, ex3)
intersectSeveral(ex1, ex4)
intersectSeveral(ex2, ex3)
intersectSeveral(ex2, ex4)
intersectSeveral(ex3, ex4)
intersectSeveral(ex1, 7)
intersectSeveral(ex3, 5)
intersectSeveral(ex3, NA)
}
