% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logindialog.R
\name{login_dialog}
\alias{login_dialog}
\title{login dialog asking for username and password
alternative to getPass() from `getPass` package}
\usage{
login_dialog(title = "", default_user = "")
}
\arguments{
\item{title}{title of the dialog}

\item{default_user}{default user name}
}
\description{
login dialog asking for username and password
alternative to getPass() from `getPass` package
}
