% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primaryKey.R
\name{lumpDuplicatedByPK}
\alias{lumpDuplicatedByPK}
\title{Lump duplicated registers by PK}
\usage{
lumpDuplicatedByPK(
  x,
  pk,
  collapse,
  tryNumeric = FALSE,
  tolerance = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{x}{a \code{data.frame}.}

\item{pk}{columns with the primary key.}

\item{collapse}{a \code{character}. If no missing, duplicated values are collapsed in a string separated by \code{collapse} character.}

\item{tryNumeric}{if \code{TRUE}, convert values to numeric if no data is lost and take the mean if values are similar.}

\item{tolerance}{relative tolerance passed to \code{all.equal} when comparing values if \code{tryNumeric} is \code{TRUE}.}
}
\value{
a named `list` with a `df` item for the x rows without duplicates and a `dup` item with duplicated rows.
}
\description{
For duplicated rows, omit \code{NA}s and take unique values in columns when at least one value exists to reduce the
duplicates.
}
