% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VbetterMax.R
\name{M2VbetterMax}
\alias{M2VbetterMax}
\title{Matrix to Vector Better Max}
\usage{
M2VbetterMax(X, MARGIN = 1)
}
\arguments{
\item{X}{a numeric matrix}

\item{MARGIN}{integer of length 1 indicating whether to get the max of each row (1) or column (2)}
}
\value{
A numeric vector where each position contains the largest value stored in that row or column of the input matrix.  If all values are NA for a particular position, that position in the output will also contain NA.
}
\description{
Collapse a 2d matrix by taking the betterMax of each row or column
}
\examples{
ex1 = matrix(c(1:5,NA,1:4,NA,2), 3, 4)
ex2 = matrix(c(5:7,NA,NA,NA), 2, 3)
M2VbetterMax(ex1, 1)
M2VbetterMax(ex1, 2)
M2VbetterMax(ex2, 1)
M2VbetterMax(ex2, 2)
}
