% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MbetterComp.R
\name{MbetterComp}
\alias{MbetterComp}
\title{Test two matrices for equality, where NA entries are considered equal}
\usage{
MbetterComp(x, y)
}
\arguments{
\item{x}{A numeric matrix}

\item{y}{A numeric matrix}
}
\value{
TRUE if all of the corresponding entries have equal values, where NA entries are considered equal.  Otherwise, FALSE.
}
\description{
Test two matrices for equality, where NA entries are considered equal
}
