% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_recursive}
\alias{merge_recursive}
\title{Merge a list of \code{data.frame}s by matching column names}
\usage{
merge_recursive(x, all = TRUE, ...)
}
\arguments{
\item{x}{a list of \code{data.frame}s.}

\item{all}{if `TRUE`, then extra rows will be added to the output, one for each row in not present in the other
´data.frame`s that has no matching rows. These rows will have NAs in those columns that are usually filled with
 values from other ´data.frame`s.}

\item{...}{other parameters to [merge()].}
}
\value{
returns a \code{data.frame} with all the merged columns and rows
}
\description{
Merge a list of \code{data.frame}s by matching column names
}
\details{
WARNING: the order of the data.frames in the list could modify the output if all=FALSE
}
