% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primaryKey.R
\name{minimumPK}
\alias{minimumPK}
\title{Find the minimum fields which make a valid Primary Key}
\usage{
minimumPK(
  x,
  fieldOrder = character(),
  excludeFields = character(),
  maxFields = ncol(x),
  minFieldSet = character()
)
}
\arguments{
\item{x}{a \code{data.frame} equivalent to a table (no duplicated registers, \code{unique(x)}).}

\item{fieldOrder}{a character vector with the sorted preferences in the fields part of the PK.}

\item{excludeFields}{columns which will be excluded from the potential primary key. Can be the index or the column
names.}

\item{maxFields}{maximum number of fields in th primary key.}

\item{minFieldSet}{columns that will be forced to be included in the primary key.}
}
\description{
Find the minimum fields which make a valid Primary Key
}
