% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreignKey.R
\name{missingFK}
\alias{missingFK}
\title{Find registers with missing Foreign Key in the related table}
\usage{
missingFK(x, y, fk, pk, searchCandidates = FALSE, ...)
}
\arguments{
\item{x}{a \code{data.frame} with a Foreign Key \code{fk} linking to \code{y}.}

\item{y}{a \code{data.frame} with a Primary Key \code{pk} linking to \code{x}.}

\item{fk}{columns containing the Foreign Key (column names or indexes) for \code{x}.}

\item{pk}{columns containing the Primary Key (column names or indexes) for \code{y}.}

\item{searchCandidates}{if \code{TRUE}, use \code{\link{agrep}} to find similar PKs for the missing FKs.}

\item{...}{parameters passed to \code{\link{agrep}}.

The order of \code{fk} and \code{pk} must match.}
}
\value{
a \code{data.frame} with the registers in \code{x} with a missing Foreign Key. If \code{searchCandidates}
  is \code{TRUE}, the result have an attribute candidates with suggested FKs.
}
\description{
Find registers with missing Foreign Key in the related table
}
