% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MessageLevel.R
\name{MLprint}
\alias{MLprint}
\title{MLprint}
\usage{
MLprint(message, ML = c(1, 1), indent = 2, indentCharacter = " ")
}
\arguments{
\item{message}{The message to be printed (or not).  It's a character vector of length 1.}

\item{ML}{integer vector length 2.  The first element indicates how deep to print messages.  The second element indicates how
many layers deep already.}

\item{indent}{integer length 1 - how many spaces to indent at each level.  Defaults to 2.}

\item{indentCharacter}{character length 1, nchar 1, the character to use for indenting messages}
}
\value{
Nothing gets returned
}
\description{
print the message level
}
\note{
This function decides whether or not to print a message, and how much to indent it.  Whenever the ML parameter is passed
  between functions, the first element should remain constant, and the second element should be increased by 1.
}
