% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{non.numeric}
\alias{non.numeric}
\title{Check non numeric values}
\usage{
non.numeric(x)
}
\arguments{
\item{x}{an atomic vector.}
}
\value{
A character vector with the values from the input vector that can't be casted to numeric.
}
\description{
Show the values that can not be casted to numeric
}
