% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primaryKey.R
\name{nonUniqueValuesByPK}
\alias{nonUniqueValuesByPK}
\title{Find values with differences for an expected unique PK value}
\usage{
nonUniqueValuesByPK(x, pk, includeNA = FALSE, aggregateFields = FALSE)
}
\arguments{
\item{x}{a \code{data.frame}.}

\item{pk}{columns with the primary key.}

\item{includeNA}{treat NA as a value or omit them}

\item{aggregateFields}{if \code{TRUE}, aggregate by fields with non-unique values instead of by PK values.}
}
\description{
Find values with differences for an expected unique PK value
}
