% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoolYear.R
\name{Oct1}
\alias{Oct1}
\title{October 1st}
\usage{
Oct1(year = schoolYear())
}
\arguments{
\item{year}{integer - the 4 digit year at the beginning of the school year.  Defaults to current school year.}
}
\value{
date - the date of October 1st day for the given school year
}
\description{
Determine the date of October 1st in a school year
}
\examples{
Oct1()
Oct1(2013)
}
