% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShortFunctions.R
\name{one.drop}
\alias{one.drop}
\title{One Drop}
\usage{
one.drop(x, keep.na = F)
}
\arguments{
\item{x}{a character vector}

\item{keep.na}{logical vector of length 1 indicating whether to keep NA elements}
}
\value{
A vector containing the elements of \code{x} that are longer than 1 character
}
\description{
Drops all elements of a character vector that have fewer than 2 characters
}
\examples{
ex1 = LETTERS[1:10]
ex2 = paste0(c("A","B","C", "D"), c("","ABC","AB","A"))
ex3 = c(NA, NA, NA, "1 2 3", "6", "591", "A ")
one.drop(ex1)
one.drop(ex2)
one.drop(ex3)
one.drop(ex3, TRUE)
}
