% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperClass.R
\name{Query}
\alias{Query}
\title{SQL Query Objects}
\usage{
Query(.x, ...)
}
\description{
Data types to represent SQL queries. It should not be necessary to use
SingleQuery and SingleQueryList interactively. \code{query} is the generic
user interface to generate SQL queries and is based on text parsing.
}
\details{
\code{SingleQuery} inherits from \code{character} and represents a single
query.

\code{SingleQueryList} inherits from \code{list} and represents a list of
single queries. It can be constructed with a list of character values.
}

