% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbind_addColumns}
\alias{rbind_addColumns}
\title{\code{rbind} adding non-matching columns filled with \code{NA}s}
\usage{
rbind_addColumns(...)
}
\arguments{
\item{...}{\code{data.frame}s.}
}
\value{
returns a \code{data.frame} with all the columns and rows
}
\description{
\code{rbind} adding non-matching columns filled with \code{NA}s
}
