% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.xslx.multi.R
\name{read.xlsx.multi}
\alias{read.xlsx.multi}
\title{Read Multiple Excel workbooks}
\usage{
read.xlsx.multi(folder, sheet = 1, pattern = NULL, dupVar = NULL,
  orderBy = NULL, decreasing = T, messageLevel = 0)
}
\arguments{
\item{folder}{character of length 1 with the address of a folder}

\item{sheet}{integer or character of length 1 indicating the worksheet to use}

\item{pattern}{character string containing a regular expression to be matched in the given character vector. Coerced by as.character to a character string if possible. If a character vector of length 2 or more is supplied, the first element is used with a warning.}

\item{dupVar}{character of length 1 with the name of a variable for which duplicates should be dropped}

\item{orderBy}{character of length 1 with the name of a variable by which to sort the result}

\item{decreasing}{logical, should ordering be done by decreasing values?}

\item{messageLevel}{integer, the level of messages that should be printed to the console}
}
\value{
data.frame
}
\description{
Read all the xlsx files in a folder that have filenames matching a given pattern and return a single data.frame
}
