% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{read_sql_file}
\alias{read_sql_file}
\title{Read SQL code from files}
\usage{
read_sql_file(fp, ...)
}
\arguments{
\item{fp}{file path}

\item{...}{other arguments to file()}
}
\description{
Read SQL code from files
}
