% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_sql_query}
\alias{read_sql_query}
\title{Send an SQL query to Athena and receive a dataframe.}
\usage{
read_sql_query(sql)
}
\arguments{
\item{sql}{An SQL query}
}
\value{
Dataframe or tibble if the tibble library is loaded.
}
\description{
Send an SQL query to Athena and receive a dataframe.
}
\examples{
`df <- dbtools::read_sql_query('select * from my_db.my_table')`
}
