% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{in_schema}
\alias{\%>\%}
\title{Package Re-exports}
\description{
These functions are used and re-exported from other packages
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dbplyr}{\code{\link[dbplyr]{in_schema}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

