% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{render_sql_template}
\alias{render_sql_template}
\title{Takes an SQL query and injects arguments with Jinja.}
\usage{
render_sql_template(sql, jinja_args)
}
\arguments{
\item{sql}{An SQL query}

\item{jinja_args}{Arguments that are referenced in the SQL file}
}
\description{
Takes an SQL query and injects arguments with Jinja.
}
