% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{repair_table}
\alias{repair_table}
\title{Run the Hive's metastore consistency check: 'MSCK REPAIR TABLE table;'.}
\usage{
repair_table(table, database)
}
\arguments{
\item{table}{Table name}

\item{database}{AWS Glue/Athena database name}
}
\value{
Query final state ('SUCCEEDED', 'FAILED', 'CANCELLED')
}
\description{
Recovers partitions and data associated with partitions.
Use this statement when you add partitions to the catalog.
It is possible it will take some time to add all partitions.
If this operation times out, it will be in an incomplete state
where only a few partitions are added to the catalog.
}
\examples{
`query_final_state = dbtools::repair_table(table='...', database='...')`
}
