% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceData.R
\name{replaceData}
\alias{replaceData}
\title{Copy data to database tables, ensuring that no primary key violation occurs}
\usage{
replaceData(db, name, value, ...)
}
\arguments{
\item{db}{one in:
\cr (\link{Credentials}) the credentials to get a connection to a database.
\cr (DBIConnection) \link[DBI]{DBIConnection-class}}

\item{name}{(character, length >= 1) data base table name(s)}

\item{value}{one in:
\cr (\link[base]{data.frame})
\cr (\link[base]{list}) of data.frames}

\item{...}{one in:
\cr for signature (Credentials, character | SingleQueryList) arguments are
passed to \code{reTry}
\cr for signature (CredentialsList) arguments are passed to the
(Credentials) method, so implicitly to reTry
\cr else ignored}
}
\description{
This functions copies data frames to database tables
}

