% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reTry.R
\name{reTry}
\alias{reTry}
\title{Re-try the call to a function}
\usage{
reTry(fun, ..., tries = 1, intSleep = 0, errorLogging = flog.error)
}
\arguments{
\item{fun}{a function}

\item{...}{arguments passed to \code{fun}}

\item{tries}{number of tries}

\item{intSleep}{interval in seconds between tries}

\item{errorLogging}{a function which is called in case of an error with
  \code{errorLogging(<try-error>, ...)}}
}
\description{
Tries to call \code{fun}. If the call results in an error, the error is
logged and after some time it is tried again.
}
\examples{
## Something that will fail every once in a while:
reTry(
  function() if (runif(1) > 0.5) stop(),
  tries = 2,
  intSleep = 1
)

## Disable logging
noErrorLogging <- function(x, ...) NULL
reTry(
  function() stop(),
  errorLogging = noErrorLogging
)
}

