% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoolYear.R
\name{schoolYear}
\alias{schoolYear}
\title{School Year}
\usage{
schoolYear(x = "year", y = Sys.Date())
}
\arguments{
\item{x}{type of information to be returned, defaults to "year".  Other acceptable values include termID, start, end, brief,
and full}

\item{y}{date, defaults to the current system date}
}
\value{
If \code{y} is year or termid, an integer is returned.  If \code{y} is start or end, a date is returned.  If \code{y}
  is brief or full, a character string is returned.
}
\description{
Determine the current school year
}
\examples{
schoolYear()
schoolYear("TermID")
schoolYear(y = as.Date("2015-09-13"))
schoolYear(x = "TermID", y = as.Date("2015-02-13"))
schoolYear(x = "end")
schoolYear("end", as.Date("2015-09-13"))
schoolYear("end", as.Date("2015-02-13"))
schoolYear(x = "start")
schoolYear("start", as.Date("2015-09-13"))
schoolYear("start", as.Date("2015-02-13"))
schoolYear("full")
schoolYear("brief")
}
