% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchFiles.R
\name{SearchFiles.intersect}
\alias{SearchFiles.intersect}
\title{Search Files Intersect}
\usage{
SearchFiles.intersect(files, include = NA_character_,
  remove = NA_character_)
}
\arguments{
\item{files}{A character vector with filenames}

\item{include}{A character vector containing regular expressions}

\item{remove}{A character vector containing regular expressions}
}
\value{
A character vector that is a subset of \code{files}.  Each element will match all elements of \code{include}, and no
  elements of the output will match any elements of \code{remove}
}
\description{
Find filenames that match ALL required regular expressions
}
