% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchFiles.R
\name{SearchFiles.union}
\alias{SearchFiles.union}
\title{Search Files Union}
\usage{
SearchFiles.union(files, include = NA_character_)
}
\arguments{
\item{files}{A character vector with filenames}

\item{include}{A character vector containing regular expressions}
}
\value{
A character vector that is a subset of \code{files}.  Each element will match at least one element of \code{include}.
}
\description{
Find filenames that match ANY required regular expressions
}
