% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendData.R
\docType{methods}
\name{sendData}
\alias{sendData}
\alias{sendData,Credentials,character,data.frame-method}
\alias{sendData,Credentials,character,list-method}
\alias{sendData,Credentials,list,list-method}
\alias{sendData,CredentialsList,character,data.frameORlist-method}
\alias{sendData,DBIConnection,character,data.frame-method}
\title{Copy data to database tables}
\usage{
sendData(db, name, value, ...)

\S4method{sendData}{CredentialsList,character,data.frameORlist}(db, name, value,
  ..., applyFun = lapply)

\S4method{sendData}{Credentials,character,data.frame}(db, name, value, ...)

\S4method{sendData}{Credentials,character,list}(db, name, value, ...)

\S4method{sendData}{Credentials,list,list}(db, name, value, ...)

\S4method{sendData}{DBIConnection,character,data.frame}(db, name, value, ...)
}
\arguments{
\item{db}{one in:
\cr (\link{Credentials}) the credentials to get a connection to a database.
\cr (DBIConnection) \link[DBI]{DBIConnection-class}}

\item{name}{(character, length >= 1) data base table name(s)}

\item{value}{one in:
\cr (\link[base]{data.frame})
\cr (\link[base]{list}) of data.frames}

\item{...}{one in:
\cr for signature (Credentials, character | SingleQueryList) arguments are
passed to \code{reTry}
\cr for signature (CredentialsList) arguments are passed to the
(Credentials) method, so implicitly to reTry
\cr else ignored}

\item{applyFun}{(function) something like lapply or mclapply}
}
\description{
This functions copies data frames to database tables
}

