% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setdiff.data.frame}
\alias{setdiff.data.frame}
\title{setdiff for data.frames with rows as a element}
\usage{
setdiff.data.frame(x, y)
}
\arguments{
\item{x}{a \code{data.frame}.}

\item{y}{a \code{data.frame}.}
}
\value{
a \code{data.frame} with the rows in \code{x} not present in \code{y}.
}
\description{
setdiff for data.frames with rows as a element
}
