% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShortFunctions.R
\name{setlength}
\alias{setlength}
\title{Set Length}
\usage{
setlength(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{integer of length 0}
}
\value{
The vector x truncated if y < length(x) or padded with NA if y > length(x)
}
\description{
Set the length of a vector
}
\examples{
ex1 = 6:10
ex2 = LETTERS[1:10]
ex3 = c(1, 10, NA, 6, 2)
ex4 = c(NA, NA, NA, 6, 5)
setlength(ex1, 7)
setlength(ex2, 5)
setlength(ex2, 15)
setlength(ex3, 4)
setlength(ex4, 50)
}
