% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setUpOracle}
\alias{setUpOracle}
\title{Verify that Oracle drive is available and sets necessary environemtn variables (Windows only)
sets env variables for Oracle driver on Windows (64-bit)}
\usage{
setUpOracle(bin_path, oci_path)
}
\arguments{
\item{bin_path}{path to Oracle binaries (such as "C:/ORACLE/11.2.0/bin")}

\item{oci_path}{path to Oracle binaries (such as "C:/ORACLE/11.2.0/oci/include")}
}
\description{
Verify that Oracle drive is available and sets necessary environemtn variables (Windows only)
sets env variables for Oracle driver on Windows (64-bit)
}
