% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{show_create_table}
\alias{show_create_table}
\title{Generate the query that created a table: 'SHOW CREATE TABLE table;'.}
\usage{
show_create_table(table, database)
}
\arguments{
\item{table}{Table name}

\item{database}{AWS Glue/Athena database name}
}
\value{
The query that created the table
}
\description{
Generate the query that created a table: 'SHOW CREATE TABLE table;'.
}
\examples{
`df_table = dbtools::show_create_table(table='my_table', database='my_db')`
}
