% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helperClass.R
\docType{class}
\name{SingleQuery-class}
\alias{SingleQuery-class}
\alias{SingleQueryList-class}
\title{Helper-Class: Single Query}
\description{
Class which represents a single query or a list of single queries. Mostly
used internally.
}
\details{
\code{SingleQuery} inherits from \code{character} and represents a single query.

\code{SingleQueryList} inherits from \code{list} and represents a list of single queries. It can be constructed with a list of character values.
}

