% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShortFunctions.R
\name{SortLength}
\alias{SortLength}
\title{Sort by Length}
\usage{
SortLength(x, decreasing = TRUE, na.last = NA)
}
\arguments{
\item{x}{a character vector}

\item{decreasing}{logical vector of length 1 indicating whether to keep NA elements.  Defaults to TRUE}

\item{na.last}{for controlling the treatment of NAs. If TRUE, missing values in the data are put last; if FALSE, they are put first; if NA, they are removed.}
}
\value{
A vector containing the elements of \code{x} sorted by length
}
\description{
Sort a character vector by the lengths of its elements
}
\examples{
ex1 = c("Green", "Tech", "High", "Charter", "School")
ex2 = c(NA, NA, NA, "1 2 3", "6", "591", "A ")
SortLength(ex1)
SortLength(ex1, F)
SortLength(ex2)
SortLength(ex2, F)
SortLength(ex2, na.last = T)
SortLength(ex2, na.last = F)
SortLength(ex2, F, na.last = T)
SortLength(ex2, F, na.last = F)
}
