% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqAssertions.R
\name{sqPattern}
\alias{sqComma}
\alias{sqEsc}
\alias{sqInNums}
\alias{sqInStrs}
\alias{sqName}
\alias{sqNames}
\alias{sqNum}
\alias{sqNums}
\alias{sqParan}
\alias{sqPattern}
\alias{sqStr}
\alias{sqStrs}
\title{SQ Assertions and Formats}
\usage{
sqPattern(x, pattern, negate = TRUE)

sqStr(x)

sqStrs(x)

sqNum(x)

sqNums(x)

sqParan(x, assert = identity)

sqComma(x, assert = identity)

sqEsc(x, assert = identity, with = "`")

sqName(x)

sqNames(x)

sqInNums(x)

sqInStrs(x)
}
\arguments{
\item{x}{(ANY) input}

\item{pattern}{(character) a regular expression used in \link{grepl}}

\item{negate}{(logical) if TRUE then an error is thrown if at least one of
the elements in x match pattern. If FALSE all elements in x must match the
pattern.}

\item{assert}{(function) an assertion fuction}
}
\description{
These functions can be used to format and check input in queries.
}

