% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql}
\alias{sql}
\title{sql}
\usage{
sql(query_string, ...)
}
\arguments{
\item{query_string}{a (character) vector or SQL to be either marked as SQL or
to be interpolated.}

\item{...}{optional furtherarguments used to interpolate values into
query_string}
}
\value{
an object of type SQL
}
\description{
sql
}
