% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_insert_conflict.R
\name{sql_conflict_insert}
\alias{sql_conflict_insert}
\title{SQL query to insert missing records}
\usage{
sql_conflict_insert(
  from,
  table,
  con,
  conflict,
  insert_cols = NULL,
  returning = NULL
)
}
\arguments{
\item{conflict}{Conflict cause generated by \code{sql_do_update()} or
\code{sql_do_nothing()}}

\item{insert_cols}{Columns from table \code{from} to insert.}
}
\description{
SQL query to insert missing records
}
\examples{
sql_conflict_insert(
  table = "my_tbl",
  con = con,
  from = "my_value_table",
  conflict = sql_do_nothing(sql_conflict_cols("conf 1", "conf 2")),
  insert_cols = c("insert 1", "insert 2"),
  returning = list(`ret 1` = "ret_col", sql("now()"))
)
}
