% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_conflict_helpers.R
\name{sql_do_nothing}
\alias{sql_do_nothing}
\title{On conflict DO NOTHING object}
\usage{
sql_do_nothing(conflict_target = NULL)
}
\arguments{
\item{conflict_target}{If \code{NULL} (the default) conflicts with all usable
constraints and unique indexes are handled. Otherwise a constraint name
given with \code{sql_constraint()} or a character vector of column names with
a unique index.}
}
\description{
On conflict DO NOTHING object
}
