% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_insert.R
\name{sql_insert}
\alias{sql_insert}
\title{SQL query to insert missing records}
\usage{
sql_insert(
  from,
  table,
  con,
  conflict = NULL,
  insert_cols = NULL,
  returning = NULL,
  return_all = FALSE
)
}
\arguments{
\item{conflict}{Conflict cause generated by \code{sql_do_update()} or
\code{sql_do_nothing()}}

\item{insert_cols}{Columns from table \code{from} to insert.}
}
\description{
SQL query to insert missing records
}
\examples{
sql_conflict_insert(
  from = "my_value_table",
  table = "my_tbl",
  con = con,
  conflict = sql_do_nothing(sql_conflict_cols("conf 1", "conf 2")),
  insert_cols = c("insert 1", "insert 2"),
  returning = list(`ret 1` = "ret_col", SQL("now()"))
)
}
