% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wip.R
\name{sql_select}
\alias{sql_select}
\title{sql_select(c(a = "a old", "b old"), con = con)
sql_select(list(a = "a old", b = sql('"table"."old"')), con = con)}
\usage{
sql_select(values, con, cols = names2(values))
}
\description{
sql_select(c(a = "a old", "b old"), con = con)
sql_select(list(a = "a old", b = sql('"table"."old"')), con = con)
}
