% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wip.R
\name{sql_set}
\alias{sql_set}
\title{sql_set(c(a = "a old", b = "b old"), con = con)
sql_set(c(a = "a old", b = "b old"), con = con, cols = c("x", "y"))
sql_set(list(a = "a old", b = sql('"table"."old"')), con = con)}
\usage{
sql_set(values, con, cols = names2(values))
}
\description{
sql_set(c(a = "a old", b = "b old"), con = con)
sql_set(c(a = "a old", b = "b old"), con = con, cols = c("x", "y"))
sql_set(list(a = "a old", b = sql('"table"."old"')), con = con)
}
