% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_update_from.R
\name{sql_update}
\alias{sql_update}
\title{SQL query to update records}
\usage{
sql_update(table, con, from, update, where, returning = NULL)
}
\arguments{
\item{table}{Name of the table to update.}

\item{con}{A DBIConnection object.}

\item{from}{Either the name of a database table (a scalar character) or
a dataframe to use for updating.}

\item{update}{Specifies which columns to update how.}

\item{where}{Expressions to use for WHERE clause
\itemize{
\item character: a column name to join by. To join by different columns on
\<table\> and \<from\> use a named vector. For example \code{where = c("a" = "b")}
will match \code{from.a} to \code{table.b}.
\item sql: must be unnamed
}}

\item{returning}{Expressions to return.
\itemize{
\item character: column to return
\item sql: expression to return
\item names: name of column in returned result.
}}
}
\value{
An SQL query.
}
\description{
SQL query to update records
}
