% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_values.R
\name{sql_values}
\alias{sql_values}
\title{sql_values}
\usage{
sql_values(x, ..., val_sep = "\\n", conn = NULL)
}
\arguments{
\item{x}{a vector or data.frame like}

\item{...}{additional parameters passed through to methods (currently not
used by standard methods)}

\item{val_sep}{a string to be put between SQL values in addition to ","}

\item{conn}{A database connection.}
}
\value{
an object of type SQL
}
\description{
A function to generate SQL value lists, to be used with sql values keyword:
  "\code{VALUES (1,2), (1,3), (2,7)}".
}
\examples{

#' # SQL VALUES list from vector
sql_values(letters)
sql_values(4L:7L)
sql_values(1.3:7.1)
sql_values(rep(Sys.time(), 10))

# SQL VALUES list from data.frame like
sql_values(data.frame(letters, seq_along(letters), Sys.time()))
sql_values(data.frame(letters, seq_along(letters), Sys.time()), conn = DBI::ANSI())


}
