% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{start_query_execution}
\alias{start_query_execution}
\title{Start a SQL Query against AWS Athena}
\usage{
start_query_execution(sql, wait = FALSE)
}
\arguments{
\item{sql}{SQL query}

\item{wait}{Default FALSE, indicates whether to wait for the query to finish and return a dictionary with the query execution response.}
}
\value{
Query execution ID if `wait` is set to `False`, list with the get_query_execution response otherwise.
}
\description{
Start a SQL Query against AWS Athena
}
\seealso{
https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/athena.html#Athena.Client.get_query_execution
}
