% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{start_query_execution_and_wait}
\alias{start_query_execution_and_wait}
\title{Calls start_query_execution followed by wait_query.}
\usage{
start_query_execution_and_wait(sql)
}
\arguments{
\item{sql}{An SQL string. Which works with __TEMP__ references.}
}
\value{
List with the get_query_execution response.
}
\description{
Calls start_query_execution followed by wait_query.
}
\examples{
`res <- dbtools::start_query_execution_and_wait('select * from __temp__.my_table')`
}
