% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{stop_query_execution}
\alias{stop_query_execution}
\title{Stop a query execution}
\usage{
stop_query_execution(query_id)
}
\arguments{
\item{query_id}{Athena query execution ID}
}
\description{
Stop a query execution
}
