% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryByPK.R
\name{summarizeByPK}
\alias{summarizeByPK}
\title{Summarize values by Primary Key}
\usage{
summarizeByPK(x, pk, collapse = "|")
}
\arguments{
\item{x}{a \code{data.frame}}

\item{pk}{columns containing the Primary Key (column names or indexes)}

\item{collapse}{characters to collapse non numeric values passed to \code{\link[base]{paste}} function}
}
\value{
a \code{data.frame} with a unique value for each PK. Mean for numeric values and a collapsed values for no numeric variables.
}
\description{
Summarize values by Primary Key
}
