% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummerSchoolFunctions.R
\name{Summer.AssignStudents}
\alias{Summer.AssignStudents}
\title{Summer Assign Students}
\usage{
Summer.AssignStudents(AssignmentList)
}
\arguments{
\item{AssignmentList}{a list containing three data.frames: SummerEnrollments, sectionTable, and students}
}
\value{
a list based on the \code{AssignmentList} input with updated versions of the three elements
}
\description{
Place students into sections of courses
}
