% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummerSchoolFunctions.R
\name{SummerSchool1}
\alias{SummerSchool1}
\title{Summer School 1}
\usage{
SummerSchool1(grades, courses, regentsScores)
}
\arguments{
\item{grades}{data.frame of grades}

\item{courses}{data.frame of course information}

\item{regentsScores}{data.frame of regents scores}
}
\value{
data.frame based on the \code{grades} input
}
\description{
Add columns to table of grades
}
