% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummerSchoolFunctions.R
\name{SummerSchool2}
\alias{SummerSchool2}
\title{Summer School 2}
\usage{
SummerSchool2(grades)
}
\arguments{
\item{grades}{data.frame of grades}
}
\value{
data.frame of unique students with counts of various different plan types
}
\description{
Create the students data.frame
}
