% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummerSchoolFunctions.R
\name{SummerSchool3}
\alias{SummerSchool3}
\title{Summer School 3}
\usage{
SummerSchool3(grades, students, maxInClass = 2, maxAllSummer = NA)
}
\arguments{
\item{grades}{data.frame of grades}

\item{students}{data.frame of students}

\item{maxInClass}{the maximum number of Summer School In Class a student can take}

\item{maxAllSummer}{the maximum number of Summer School In Class or Summer On Plato courses (this currently does nothing)}
}
\value{
data.frame based on the \code{grades} input
}
\description{
Resolve most Possibly Summer situations
}
