% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummerSchoolFunctions.R
\name{SummerSchool4}
\alias{SummerSchool4}
\title{Summer School 4}
\usage{
SummerSchool4(courses, grades)
}
\arguments{
\item{courses}{data.frame of course information}

\item{grades}{data.frame of grades}
}
\value{
data.frame of summer courses
}
\description{
Determine the Summer Courses
}
