% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummerSchoolFunctions.R
\name{SummerSchool5}
\alias{SummerSchool5}
\title{Summer School 5}
\usage{
SummerSchool5(students, RevisedGrades)
}
\arguments{
\item{students}{data.frame of students}

\item{RevisedGrades}{data.frame of grades}
}
\value{
data.frame based on the \code{students} input
}
\description{
Add plan counts to the students data.frame
}
