% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummerSchoolFunctions.R
\name{SummerSchool6}
\alias{SummerSchool6}
\title{Summer School 6}
\usage{
SummerSchool6(RevisedGrades, SummerCourses)
}
\arguments{
\item{RevisedGrades}{data.frame of grades}

\item{SummerCourses}{data.frame of summer courses}
}
\value{
data.frame based on the \code{RevisedGrades} input
}
\description{
If a course is canceled due to underenrollment, reassign the student
}
