% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShortFunctions.R
\name{SWSM}
\alias{SWSM}
\title{Suppress Warnings and Messages}
\usage{
SWSM(expr)
}
\arguments{
\item{expr}{expression to evaluate}
}
\description{
Suppress warnings and messages when evaluating an expression
}
