% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testConnection.R
\docType{methods}
\name{testConnection}
\alias{testConnection}
\alias{testConnection,Credentials-method}
\alias{testConnection,CredentialsList-method}
\title{testConnection}
\usage{
testConnection(x, logging = TRUE, ...)

\S4method{testConnection}{Credentials}(x, logging = TRUE, ...)

\S4method{testConnection}{CredentialsList}(x, logging = TRUE, ...)
}
\arguments{
\item{x}{(Credentials | CredentialsList)}

\item{logging}{(logical)}

\item{...}{arguments passed to \link{sendQuery}}
}
\description{
testConnection
}

