% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testConnection.R
\docType{methods}
\name{testConnection}
\alias{loggerConnection}
\alias{loggerSuppress}
\alias{testConnection}
\alias{testConnection,Credentials-method}
\alias{testConnection,CredentialsList-method}
\title{Test a connection}
\usage{
testConnection(x, logger = loggerConnection, ...)

\S4method{testConnection}{Credentials}(x, logger = loggerConnection, ...)

\S4method{testConnection}{CredentialsList}(x, logger = loggerConnection, ...)

loggerConnection(x, status)

loggerSuppress(x, status)
}
\arguments{
\item{x}{(Credentials | CredentialsList) a credentials object.}

\item{logger, status}{(function) a logger function. Has two arguments, first is
\code{x} second is used to indicate success and failure as
\code{logical(1)}.}

\item{...}{arguments passed to \link{sendQuery}}
}
\description{
Test your conncection to local or remote servers. The function will return a
logical of length one invisibly with \code{TRUE} for success and \code{FALSE}
if any of the connection attempts fail.
}
\examples{

workingConnection <- Credentials(drv = RSQLite::SQLite, dbname = ":memory:")
testConnection(workingConnection)

## To suppress logging:
testConnection(workingConnection, loggerSuppress)

}

