% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnEnrolledInCohort.R
\name{UnEnrolledInCohort}
\alias{UnEnrolledInCohort}
\title{Unenrolled Students In The Cohort}
\usage{
UnEnrolledInCohort(
  WB,
  cohorts,
  returnvars = c("Local.ID.(optional)", "Last.Name", "First.Name",
    "Cohort.Year.(year.1st.entered.9th)", "Date.left.GTH", "Discharge.Reason"),
  messageLevel = 0
)
}
\arguments{
\item{WB}{data.frame of the accountability workbook}

\item{cohorts}{integer - how many cohorts should be analyzed?}

\item{returnvars}{character vector with the names of the columns to be included in the return}

\item{messageLevel}{integer - what degree of messaging should be printed to the console?}
}
\value{
a data.frame with all of the students who are in their grad cohort but not enrolled
}
\description{
Finds all students who are no longer enrolled but still in the graduation cohort
}
