% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateDescription.R
\name{UpdateDescription}
\alias{UpdateDescription}
\title{Update Description}
\usage{
UpdateDescription(date = Sys.Date(), path = getwd())
}
\arguments{
\item{date}{Date object, defaults to current date}

\item{path}{File path, defaults to working directory}
}
\value{
Nothing is returned
}
\description{
Update the Date and Version fields in the DESCRIPTION file
}
