% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterGreater.R
\name{VbetterGreater}
\alias{VbetterGreater}
\title{Vector Better Greater Than}
\usage{
VbetterGreater(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector of the same length as \code{x} or of length 1 (if y has length 1, it will be repeated to the length of x)}
}
\value{
A logical vector where each position indicates whether the values in those positions of the first input vector is greater than that of the second. NAs are considered equal to each other, and less than -Inf.
}
\description{
Compare two vectors for element by element to see if the first is greater than the second
}
\examples{
ex1 = 6:10
ex2 = (1:5)*2
ex3 = c(1, 10, NA, 6, 10)
ex4 = c(2, NA, NA, 6, 5)
VbetterGreater(ex1, ex2)
VbetterGreater(ex1, ex3)
VbetterGreater(ex1, ex4)
VbetterGreater(ex2, ex3)
VbetterGreater(ex2, ex4)
VbetterGreater(ex3, ex4)
VbetterGreater(ex3, NA)
VbetterGreater(ex3, 10)
}
