% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterGrepl.R
\name{VbetterGrepl}
\alias{VbetterGrepl}
\title{Vector Better grepl}
\usage{
VbetterGrepl(pattern, x, ignore.case = T, fixed = T)
}
\arguments{
\item{pattern}{character vector containing regular expressions (or character strings for fixed = TRUE) to be matched in the
given character vector \code{x}. Coerced by as.character to a character string if possible.}

\item{x}{a character vector where matches are sought, or an object which can be coerced by as.character to a character vector.
Long vectors are supported.}

\item{ignore.case}{logical.  if FALSE, the pattern matching is case sensitive and if TRUE, case is ignored during matching.}

\item{fixed}{logical. If TRUE, \code{pattern}s are strings to be matched as is. Overrides all conflicting arguments.}
}
\value{
Logical matrix with one row for each element of \code{x} and one column for each element of \code{pattern}.  The
  entries indicate whether a math was found between the relevant pattern and character string.
}
\description{
This function behaves similarly to betterGrepl, except that it accepts a vector of \code{pattern}s
}
\examples{
x = c("Alpha", "Beta", "Bet", NA_character_, "B")
patterns = c("a", NA, "b", "c", "e")
VbetterGrepl(pattern = patterns, x = x)
VbetterGrepl(pattern = patterns, x = x, ignore.case = F)
VbetterGrepl(pattern = patterns, x = x, fixed = F)
VbetterGrepl(pattern = patterns, x = x, ignore.case = F, fixed = F)
}
