% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geocode.R
\name{VgetSchoolDistrict}
\alias{VgetSchoolDistrict}
\title{Vector Get School District}
\usage{
VgetSchoolDistrict(addrVec, halt = T, na.rm = F, messageLevel = 0)
}
\arguments{
\item{addrVec}{Character vector containing addresses}

\item{halt}{logical of length 1 indicating whether to halt if no school district is found}

\item{na.rm}{logical of length 1 indicating whether to remove NA's from the returned vector (ignored if \code{halt} is TRUE)}

\item{messageLevel}{integer of length 1 indicating the level of messaging to print}
}
\value{
Character vector containing the names of the school districts or NA's
}
\description{
Takes a vector of one line addresses (USA only) and returns the school district serving those address using the first matches from the US Census API
}
\examples{
addrVecExample = c("1 MAIN ST, ALBANY, NY, 12204", "1234 MAIN Blvd, ALBANY, NY, 12204", "1234 MAIN Blvd, ALBANY, NY, 12204")
VgetSchoolDistrict(addrVecExample)
VgetSchoolDistrict(addrVecExample, halt = F)
VgetSchoolDistrict(addrVecExample, halt = F, na.rm = T)
}
