% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MessageLevel.R
\name{vMLprint}
\alias{vMLprint}
\title{Vector MLprint}
\usage{
vMLprint(messages, ML = c(0, 0), indent = 2, indentCharacter = " ")
}
\arguments{
\item{messages}{The messages to be printed (or not).  Character vector where each element is one message to print.}

\item{ML}{integer vector length 2.  The first element indicates how deep to print messages.  The second element indicates how
many layers deep already.}

\item{indent}{integer length 1 - how many spaces to indent at each level.  Defaults to 2.}

\item{indentCharacter}{character length 1, nchar 1, the character to use for indenting messages}
}
\value{
Nothing gets returned
}
\description{
print the message level
}
\note{
This function acts as a vectorized version of MLprint
}
