% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{wait_query}
\alias{wait_query}
\title{Wait for a query to end.}
\usage{
wait_query(query_id)
}
\arguments{
\item{query_id}{Athena query execution ID}
}
\value{
List with the get_query_execution response
}
\description{
Wait for a query to end.
}
\examples{
`res <- dbtools::wait_query(query_id)`
}
