% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbt_get_query.R
\name{wpd_get_query}
\alias{wpd_get_query}
\title{wpd_get_query}
\usage{
wpd_get_query(query, verbose = TRUE, con = NULL, node = NULL,
  close = NULL, dt = TRUE)
}
\arguments{
\item{query}{SQL statement to execute}

\item{verbose}{return extra information -- timings, potential database errors -- or not}
}
\description{
Establish a connection, execute SQL statement against database, close connection and return.
}
\examples{


}
