% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WritePSimport.R
\name{write.psimport}
\alias{write.psimport}
\title{Write PowerSchool Import File}
\usage{
write.psimport(x, file, messageLevel = 0)
}
\arguments{
\item{x}{the object to be written, preferably a matrix or data frame. If not, it is attempted to coerce x to a data frame.}

\item{file}{a character string naming a file}

\item{messageLevel}{integer of length 1 indicating the level of messaging to print}
}
\value{
Nothing
}
\description{
This is a wrapper for write.table that produces a file formatted to be imported into PowerSchool
}
