% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShortFunctions.R
\name{xlDate}
\alias{xlDate}
\title{Convert dates from Excel}
\usage{
xlDate(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
A vector containing the elements of \code{x} converted to Date format
}
\description{
Convert date values pulled from an Excel workbook, such as .xslx.  This function just runs as.Date with the origin set to "1899-12-30".  Excel apparently uses this as the origin because they did not realize that 1900 was not a leap year.
}
\examples{
xlDate(42359)
xlDate(0)
xlDate(-1365)
datevector = 40000 + (1:10)*365
xlDate(datevector)
}
