% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{convert_athena_type_to_arrow}
\alias{convert_athena_type_to_arrow}
\title{Convert an Athena type to an Arrow type}
\usage{
convert_athena_type_to_arrow(t)
}
\arguments{
\item{t}{A string giving an Athena type}
}
\value{
An Arrow type
}
\description{
Convert an Athena type to an Arrow type
}
\seealso{
https://docs.aws.amazon.com/athena/latest/ug/data-types.html

https://arrow.apache.org/docs/r/reference/data-type.html
}
