% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Credentials.R
\docType{class}
\name{Credentials-class}
\alias{Credentials}
\alias{Credentials-class}
\alias{CredentialsList}
\alias{CredentialsList-class}
\alias{as.list.Credentials}
\title{Credentials Class}
\usage{
Credentials(drv, ...)

\method{as.list}{Credentials}(x, ...)

CredentialsList(...)
}
\arguments{
\item{drv}{(function) a driver. Will be called and passed to \link[DBI]{dbConnect}}

\item{...}{named arguments which can be passed to \link[DBI]{dbConnect}}

\item{x}{(Credentials) an instance}
}
\description{
Use this function to store your database credentials.
}
\details{
\code{CredentialsList} can be used to construct a list of Credential
  objects. The advantage is, that all arguments can be vectors. Elements of
  length one are replicated to match the appropriate number of credentials.
  This is usefull whenever you run the same query on multiple databases where
  they only differ in the port but else expect the same credentials.
}
\examples{
Credentials(drv = RSQLite::SQLite, dbname = ":memory:")
Credentials(drv = RSQLite::SQLite, dbname = c(":memory:", ":memory:"))
}

