% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-column-binary.R
\name{db_column_binary}
\alias{db_column_binary}
\title{Binary Column}
\usage{
db_column_binary(
  table,
  name,
  default = NULL,
  validate = function(value) { },
  nullable = TRUE
)
}
\description{
Binary Column
}
