% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-column-integer.R
\name{db_column_integer}
\alias{db_column_integer}
\title{Integer Column}
\usage{
db_column_integer(
  name,
  default = NULL,
  intrange = 2147483647L,
  validate = function(value) { },
  nullable = TRUE,
  identity = c("none", "default", "always"),
  .x = list(),
  .class = character()
)
}
\arguments{
\item{identity}{is this an identity column? Can be:
\itemize{
\item \code{"none"}    - not an identity column
\item \code{"default"} - can be overwritten on insert
\item \code{"always"}  - can not be overwritten
}}

\item{integer}{integers will be bound within the range \link{-intrange, +intrange}}
}
\description{
Integer Column
}
