% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-column-text.R
\name{db_column_text}
\alias{db_column_text}
\title{Text Column}
\usage{
db_column_text(
  name,
  default = NULL,
  max_char = Inf,
  validate = function(value) { },
  nullable = TRUE,
  .x = list(),
  .class = character()
)
}
\arguments{
\item{max_char}{maximum length}
}
\description{
Text Column
}
