% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_update_data.R
\name{db_update_data}
\alias{db_update_data}
\title{Update a database table}
\usage{
db_update_data()
}
\value{
If \code{returning} is \code{NULL} the number of rows updated; Otherwise a
tibble of the updated rows and the columns as specified in \code{returning}.
}
\description{
Update a database table
}
