% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_upsert_data.R
\name{db_upsert_data}
\alias{db_upsert_data}
\title{Update a database table}
\usage{
db_upsert_data(
  data,
  table,
  con,
  update,
  conflict_target,
  insert_cols = NULL,
  returning = NULL,
  trans = TRUE,
  batch_size = 50000
)
}
\value{
If \code{returning} is \code{NULL} the number of rows updated; Otherwise a
tibble of the updated rows and the columns as specified in \code{returning}..
}
\description{
Update a database table
}
\section{Unnecessary Updates}{

Be careful when doing updates very frequently on a table. On an update
operation Postgres writes a new version of a row even when nothing has
changed (due to its \href{https://www.postgresql.org/docs/current/mvcc-intro.html}{MVCC model}).
This means that frequently updating a table without changing a value can
still bloat the table. For example this can easily happen together when
updating a timestamp column via a trigger.
}

