% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTools-package.R
\docType{package}
\name{dbTools-package}
\alias{dbTools}
\alias{dbTools-package}
\title{dbTools: Tools for relational databases in R}
\description{
Functions for integrity check and fixes for primary keys and foreign keys for tables in a relational DB.
}
\author{
\strong{Maintainer}: Joan Maspons \email{j.maspons@creaf.uab.cat} (\href{https://orcid.org/0000-0003-2286-8727}{ORCID})

}
\keyword{internal}
