% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbtools.R
\docType{package}
\name{dbtools-package}
\alias{dbtools}
\alias{dbtools-package}
\title{dbtools: A toolbox to work with databases}
\description{
Provides helpers for insert, update and upsert operations in databases.
}
\author{
\strong{Maintainer}: Maximilian Girlich \email{maximilian.girlich@metoda.com}

}
