% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_insert.R
\name{default-args}
\alias{default-args}
\title{Common arguments}
\arguments{
\item{data}{A data frame or the name of a database table.}

\item{table}{The name of the database table to update, insert to, or upsert
to. Can be an unquoted character string or an identifier created with \code{\link[=ident]{ident()}}.}

\item{con}{A DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}
}
\description{
Common arguments
}
