% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natoempty.R
\name{DFna.to.empty}
\alias{DFna.to.empty}
\title{Data Frame NA to Empty}
\usage{
DFna.to.empty(x, coltypes = "both")
}
\arguments{
\item{x}{A data.frame}

\item{coltypes}{Character of length 1 indicating what types of columns to affect.  Acceptable values are "character", "factor", and "both".}
}
\value{
A data.frame with the same dimensions and names as \code{x}
}
\description{
Converts all NA's in the character and/or factor columns of a data.frame vector to empty values
}
\examples{
V1 = c(LETTERS[1:8], NA, NA)
V2 = c(NA, 1:8, NA)
V3 = factor(c(NA,rep(c("A","B"), times = 4),NA))
ex1 = data.frame(V1 = V1, V2 = V2, V3 = V3, stringsAsFactors = F)
DFna.to.empty(ex1)
}
