% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MbetterComp.R
\name{MbetterComp}
\alias{MbetterComp}
\title{Test two matrices for equality, where NA entries are considered equal}
\usage{
MbetterComp(x, y)
}
\arguments{
\item{x}{A numeric matrix}

\item{y}{A numeric matrix}
}
\value{
TRUE if all of the corresponding entries have equal values, where NA entries are considered equal.  Otherwise, FALSE.
}
\description{
Test two matrices for equality, where NA entries are considered equal
}
\examples{
m1 = matrix(data = c(1:12,1:12), nrow = 6)
m2 = matrix(1:24,nrow = 6)
m3 = matrix(1:24,nrow = 4)
m4 = matrix(c(rep(NA,12),1:12)
MbetterComp(m1, m2)
MbetterComp(m1, m3)
MbetterComp(m1, m4)
MbetterComp(m2, m3)
MbetterComp(m2, m4)
MbetterComp(m3, m4)
}
