% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natoempty.R
\name{na.to.empty}
\alias{na.to.empty}
\title{NA to Empty}
\usage{
na.to.empty(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A numeric vector where each position is the largest value stored in those positions of the input vectors.  If both input vectors have NA for a particular position, that position in the output will also contain NA.
}
\description{
Converts all NA's in a character vector to empty values
}
\examples{
ex1 = LETTERS[1:10]
ex2 = c(NA, letters, NA)
ex3 = c(NA, NA, NA, NA, "A")
na.to.empty(ex1)
na.to.empty(ex2)
na.to.empty(ex3)
}
