% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\docType{class}
\name{Query}
\alias{Query}
\alias{SingleQuery-class}
\alias{SingleQueryList-class}
\title{SQL Query Objects}
\usage{
Query(.x, ..., .data = NULL, .envir = parent.frame(),
  checkSemicolon = TRUE)
}
\arguments{
\item{.x}{(character | connection) A character or connection containing a
query.}

\item{...}{Parameters to be substituted in .x}

\item{.data}{(list)}

\item{.envir}{(environment) Should be left with the default. Sets the
environment in which to evaluate code chunks in queries.}
}
\description{
Data types to represent SQL queries. It should not be necessary to use
SingleQuery and SingleQueryList interactively. \code{Query} is the generic
user interface to generate SQL queries and is based on text parsing.
}
\details{
\code{SingleQuery} inherits from \code{character} and represents a single
query.

\code{SingleQueryList} inherits from \code{list} and represents a list of
single queries. It can be constructed with a list of character values.
}
\examples{
query1 <- "SELECT {{ varName }} FROM {{ tableName }} WHERE primaryKey = {{ id }};"
query2 <- "SHOW TABLES;"

Query(query1, varName = "someVar", tableName = "someTable", .data = list(id = 1:2))

tmpFile <- tempfile()
writeLines(c(query1, query2), tmpFile)
Query(file(tmpFile))
}
