% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_sql_query}
\alias{read_sql_query}
\title{Send an SQL query to Athena and receive a data frame.}
\usage{
read_sql_query(sql)
}
\arguments{
\item{sql}{An SQL query}
}
\value{
Dataframe
}
\description{
Send an SQL query to Athena and receive a data frame.
}
