% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbtools.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{glue_sql}
\alias{SQL}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{DBI}{\code{\link[DBI]{SQL}}}

  \item{glue}{\code{\link[glue]{glue_sql}}}
}}

