% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoolYear.R
\name{schoolYear}
\alias{schoolYear}
\title{School Year}
\usage{
schoolYear(x = "year", y = Sys.Date())
}
\arguments{
\item{x}{type of information to be returned, defaults to "year"}

\item{y}{date, defaults to the current system date}
}
\value{
integer - the year or term id
}
\description{
Determine the current school year
}
\examples{
schoolYear()
schoolYear("TermID")
schoolYear(y = as.Date("2015-09-13"))
schoolYear(x = "TermID", y = as.Date("2015-02-13"))
}
