% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoolYear.R
\name{schoolYear}
\alias{schoolYear}
\title{School Year}
\usage{
schoolYear(x = "year", y = Sys.Date())
}
\arguments{
\item{x}{type of information to be returned, defaults to "year"}

\item{y}{date, defaults to the current system date}
}
\value{
TRUE if the two values are equal or if both are NA, and FALSE if they are not equal or if exactly one is NA
}
\description{
Determine the current school year
}
\note{
Instead of two arguments, this function will also accept a single vector of length 2 and compare the the two elements.
}
\examples{
schoolYear()
schoolYear("TermID")
schoolYear(y = as.Date("2015-09-13"))
schoolYear(x = "TermID", y = as.Date("2015-02-13"))
}
