% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sendQuery.R
\name{sendQuery}
\alias{sendQuery}
\alias{sendQuery.DBCredentials}
\alias{sendQuery.DBIConnection}
\alias{sendQuery.MySQLConnection}
\alias{sendQuery.list}
\title{Send query to database and fetch result}
\usage{
sendQuery(db, query, ...)

\method{sendQuery}{DBCredentials}(db, query, ...)

\method{sendQuery}{list}(db, query, ...)

\method{sendQuery}{DBIConnection}(db, query, ...)

\method{sendQuery}{MySQLConnection}(db, query, ...)
}
\arguments{
\item{db}{a database}

\item{query}{a query}

\item{...}{arguments passed to \code{reTry}}
}
\description{
This functions sends a query to a database and fetches the result.
}

