% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sendQuery.R
\docType{methods}
\name{sendQuery}
\alias{sendQuery}
\alias{sendQuery,ArgList,SingleQuery-method}
\alias{sendQuery,Credentials,character-method}
\alias{sendQuery,DBIConnection,SingleQuery-method}
\alias{sendQuery,MySQLConnection,SingleQuery-method}
\title{Send query to database and fetch result}
\usage{
sendQuery(db, query, ...)

\S4method{sendQuery}{Credentials,character}(db, query, ...)

\S4method{sendQuery}{ArgList,SingleQuery}(db, query, ...)

\S4method{sendQuery}{DBIConnection,SingleQuery}(db, query, ...)

\S4method{sendQuery}{MySQLConnection,SingleQuery}(db, query, ...)
}
\arguments{
\item{db}{(Credentials) the credentials to get a connection to a database.}

\item{query}{(character, length >= 1) a query.}

\item{...}{arguments passed to \code{reTry}}
}
\description{
This functions sends a query to a database and fetches the result.
}

