% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqAssertions.R
\name{sqPattern}
\alias{sqChar}
\alias{sqChars}
\alias{sqComma}
\alias{sqEsc}
\alias{sqInNums}
\alias{sqInStrs}
\alias{sqName}
\alias{sqNames}
\alias{sqNum}
\alias{sqNums}
\alias{sqParan}
\alias{sqPattern}
\title{SQ Assertions and Formats}
\usage{
sqPattern(x, pattern, negate = TRUE)

sqChar(x)

sqChars(x)

sqNum(x)

sqNums(x)

sqParan(x, assert = identity)

sqComma(x, assert = identity)

sqEsc(x, assert = identity, with = "`")

sqName(x)

sqNames(x)

sqInNums(x)

sqInStrs(x)
}
\arguments{
\item{x}{(ANY) input}

\item{pattern}{(character) a regular expression used in \link{grepl}}

\item{negate}{(logical) if TRUE then an error is thrown if at least one of
the elements in x match pattern. If FALSE all elements in x must match the
pattern.}

\item{assert}{(function) an assertion fuction}

\item{with}{(character)}
}
\description{
These functions can be used to format and check the input of queries.
}
\examples{
# Will format and check:
sqInStrs(letters[1:2])
sqInNums(1:2)
sqNames(letters[1:2])
sqNames("a")

# Only check:
sqNum(1)
sqNums(1:2)
sqChar("a")
sqChars(letters[1:2])
}

