% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_delete.R
\name{sql_delete}
\alias{sql_delete}
\title{SQL query to delete records}
\usage{
sql_delete(from, table, con, where, returning = NULL)
}
\description{
SQL query to delete records
}
\examples{
sql_delete(
  table = "my_tbl",
  con = con,
  from = "my_value_table",
  where = list("where 1", sql("my_tbl.id > 1")),
  returning = list(`ret 1` = "ret_col", sql("now()"))
)
}
