% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_conflict_helpers.R
\name{sql_do_nothing}
\alias{sql_do_nothing}
\title{On conflict DO NOTHING object}
\usage{
sql_do_nothing(conflict_target)
}
\arguments{
\item{conflict_target}{specifies the conflict target. This can be one of
the following:
\itemize{
\item a character vector of column names that should be unique.
When using this together with mode "new" there must be a unique constraint
on these columns.
\item a constraint name specified by \code{\link[=sql_constraint]{sql_constraint()}}.
}}
}
\description{
On conflict DO NOTHING object
}
