% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_conflict_helpers.R
\name{sql_do_update}
\alias{sql_do_update}
\title{On conflict DO UPDATE object}
\usage{
sql_do_update(conflict_target, update)
}
\arguments{
\item{conflict_target}{A constraint name given with \code{sql_constraint()} or
a character vector of column names with a unique index.}
}
\description{
On conflict DO UPDATE object
}
