% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_insert_conflict.R
\name{sql_insert_missing}
\alias{sql_insert_missing}
\title{SQL for inserting missing values}
\usage{
sql_insert_missing(
  from,
  table,
  con,
  conflict_target = NULL,
  insert_cols = NULL,
  returning = NULL,
  return_all = FALSE
)
}
\description{
SQL for inserting missing values
}
\examples{
DBI::dbWriteTable(con, "dbtools_mtcars", mtcars[1, ], overwrite = TRUE)
DBI::dbExecute(
  con,
  'ALTER TABLE dbtools_mtcars
  ADD CONSTRAINT unique_const
  UNIQUE ("mpg", "cyl");'
)
sql_insert_missing(
  from = mtcars[c(1, 3), ],
  table = "dbtools_mtcars",
  con = con,
  conflict_target = sql_conflict_cols("mpg", "cyl"),
  returning = sql("*"),
  return_all = TRUE
)
}
