% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_update_from.R
\name{sql_update}
\alias{sql_update}
\title{SQL query to update records}
\usage{
sql_update(data, table, con, update, where, returning = NULL)
}
\arguments{
\item{data}{A data.frame.}

\item{table}{Name of the table to update.}

\item{con}{A DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{update}{Specifies which columns to update how.}

\item{where}{Expressions to use for WHERE clause
\itemize{
\item character: a column name to join by. To join by different columns on
\<table\> and \<from\> use a named vector. For example \code{where = c("a" = "b")}
will match \code{from.a} to \code{table.b}.
\item sql: must be unnamed
}}

\item{returning}{An \href{sql-expression-list}{SQL expression list} that specifies
which columns to return. Names are optional and used as column names in
the returned table. For the default \code{NULL} no rows are returned.}
}
\value{
An SQL query.
}
\description{
SQL query to update records
}
