% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VbetterComp.R
\name{VbetterComp}
\alias{VbetterComp}
\title{Vector Better Comparison}
\usage{
VbetterComp(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector of the same length as \code{x} or of length 1 (if y has length 1, it will be repeated to the length of x)}
}
\value{
A logical vector where each position indicates whether the values in those positions of the input vectors are equal.  NAs are considered equal.
}
\description{
Compare two vectors for equality element by element
}
\examples{
ex1 = 6:10
ex2 = (1:5)*2
ex3 = c(1, 10, NA, 6, 10)
ex4 = c(2, NA, NA, 6, 5)
VbetterComp(ex1, ex2)
VbetterComp(ex1, ex3)
VbetterComp(ex1, ex4)
VbetterComp(ex2, ex3)
VbetterComp(ex2, ex4)
VbetterComp(ex3, ex4)
VbetterComp(ex3, NA)
VbetterComp(ex3, 10)
}
